+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/autoinstall/fragment/FragmentInstallStrategy.kt
@@ -11,13 +11,20 @@ import org.slf4j.Logger
 // @CacheableRule // TODO: make it cacheable somehow (probably depends on parameters)
 abstract class FragmentInstallStrategy : AbstractInstallStrategy {
 
-  constructor(logger: Logger) : super() {
+  constructor(
+    autoInstallEnabled: Boolean,
+    sentryVersion: String,
+    logger: Logger,
+  ) : super(autoInstallEnabled, sentryVersion) {
     this.logger = logger
   }
 
   @Suppress("unused") // used by Gradle
   @Inject // inject is needed to avoid Gradle error
-  constructor() : this(SentryPlugin.logger)
+  constructor(
+    autoInstallEnabled: Boolean,
+    sentryVersion: String,
+  ) : this(autoInstallEnabled, sentryVersion, SentryPlugin.logger)
 
   override val sentryModuleId: String
     get() = SENTRY_FRAGMENT_ID
@@ -30,8 +37,14 @@ abstract class FragmentInstallStrategy : AbstractInstallStrategy {
     private const val FRAGMENT_ID = "fragment"
     internal const val SENTRY_FRAGMENT_ID = "sentry-android-fragment"
 
-    override fun register(component: ComponentMetadataHandler) {
-      component.withModule("$FRAGMENT_GROUP:$FRAGMENT_ID", FragmentInstallStrategy::class.java) {}
+    override fun register(
+      component: ComponentMetadataHandler,
+      autoInstallEnabled: Boolean,
+      sentryVersion: String,
+    ) {
+      component.withModule("$FRAGMENT_GROUP:$FRAGMENT_ID", FragmentInstallStrategy::class.java) {
+        it.params(autoInstallEnabled, sentryVersion)
+      }
     }
   }
 }