+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/autoinstall/kotlin/KotlinExtensionsInstallStrategy.kt
@@ -11,13 +11,20 @@ import org.slf4j.Logger
 // @CacheableRule
 abstract class KotlinExtensionsInstallStrategy : AbstractInstallStrategy {
 
-  constructor(logger: Logger) : super() {
+  constructor(
+    autoInstallEnabled: Boolean,
+    sentryVersion: String,
+    logger: Logger,
+  ) : super(autoInstallEnabled, sentryVersion) {
     this.logger = logger
   }
 
   @Suppress("unused") // used by Gradle
   @Inject // inject is needed to avoid Gradle error
-  constructor() : this(SentryPlugin.logger)
+  constructor(
+    autoInstallEnabled: Boolean,
+    sentryVersion: String,
+  ) : this(autoInstallEnabled, sentryVersion, SentryPlugin.logger)
 
   override val sentryModuleId: String
     get() = SENTRY_KOTLIN_EXTENSIONS_ID
@@ -37,11 +44,17 @@ abstract class KotlinExtensionsInstallStrategy : AbstractInstallStrategy {
 
     private val MIN_SUPPORTED_VERSION = SemVer(1, 6, 1)
 
-    override fun register(component: ComponentMetadataHandler) {
+    override fun register(
+      component: ComponentMetadataHandler,
+      autoInstallEnabled: Boolean,
+      sentryVersion: String,
+    ) {
       component.withModule(
         "$KOTLINX_GROUP:$KOTLIN_COROUTINES_ID",
         KotlinExtensionsInstallStrategy::class.java,
-      ) {}
+      ) {
+        it.params(autoInstallEnabled, sentryVersion)
+      }
     }
   }
 }