+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/autoinstall/spring/SpringBoot4InstallStrategy.kt
@@ -11,13 +11,20 @@ import org.slf4j.Logger
 // @CacheableRule
 abstract class SpringBoot4InstallStrategy : AbstractInstallStrategy {
 
-  constructor(logger: Logger) : super() {
+  constructor(
+    autoInstallEnabled: Boolean,
+    sentryVersion: String,
+    logger: Logger,
+  ) : super(autoInstallEnabled, sentryVersion) {
     this.logger = logger
   }
 
   @Suppress("unused") // used by Gradle
   @Inject // inject is needed to avoid Gradle error
-  constructor() : this(SentryPlugin.logger)
+  constructor(
+    autoInstallEnabled: Boolean,
+    sentryVersion: String,
+  ) : this(autoInstallEnabled, sentryVersion, SentryPlugin.logger)
 
   override val sentryModuleId: String
     get() = SENTRY_SPRING_BOOT_4_ID
@@ -39,11 +46,17 @@ abstract class SpringBoot4InstallStrategy : AbstractInstallStrategy {
     private val MIN_SUPPORTED_VERSION = SemVer(4, 0, 0, "M1")
     private val MAX_SUPPORTED_VERSION = SemVer(4, 9999, 9999)
 
-    override fun register(component: ComponentMetadataHandler) {
+    override fun register(
+      component: ComponentMetadataHandler,
+      autoInstallEnabled: Boolean,
+      sentryVersion: String,
+    ) {
       component.withModule(
         "$SPRING_GROUP:$SPRING_BOOT_4_ID",
         SpringBoot4InstallStrategy::class.java,
-      ) {}
+      ) {
+        it.params(autoInstallEnabled, sentryVersion)
+      }
     }
   }
 }